centralizing eventbus related interfaces and structs into the eventbus directory
