+++ eventbus/jetstream/sensor/sensorJetstream.go
@@ -1,4 +1,4 @@
-package sensoreventbus
+package sensor
 
 import (
 	"fmt"
@@ -6,7 +6,8 @@ import (
 
 	"math/rand"
 
-	eventbusdriver "github.com/argoproj/argo-events/eventbus/driver"
+	eventbuscommon "github.com/argoproj/argo-events/eventbus/common"
+	eventbusjetstreambase "github.com/argoproj/argo-events/eventbus/jetstream/base"
 	"github.com/argoproj/argo-events/pkg/apis/sensor/v1alpha1"
 	nats "github.com/nats-io/nats.go"
 	"go.uber.org/zap"
@@ -14,16 +15,16 @@ import (
 	"github.com/argoproj/argo-events/common"
 )
 
-type Jetstream struct {
-	*eventbusdriver.Jetstream
+type SensorJetstream struct {
+	*eventbusjetstreambase.Jetstream
 
 	sensorName    string
 	keyValueStore nats.KeyValue
 }
 
-func NewJetstream(url string, sensorSpec *v1alpha1.Sensor, auth *eventbusdriver.Auth, logger *zap.SugaredLogger) (*Jetstream, error) {
+func NewSensorJetstream(url string, sensorSpec *v1alpha1.Sensor, auth *eventbuscommon.Auth, logger *zap.SugaredLogger) (*SensorJetstream, error) {
 
-	baseJetstream, err := eventbusdriver.NewJetstream(url, auth, logger)
+	baseJetstream, err := eventbusjetstreambase.NewJetstream(url, auth, logger)
 	if err != nil {
 		return nil, err
 	}
@@ -40,14 +41,14 @@ func NewJetstream(url string, sensorSpec *v1alpha1.Sensor, auth *eventbusdriver.
 	// todo: Here we can take the sensor specification and clean up the K/V store so as to remove any old
 	// Triggers for this Sensor that no longer exist and any old Dependencies (and also Drain any corresponding Connections)
 
-	return &Jetstream{
+	return &SensorJetstream{
 		baseJetstream,
 		sensorSpec.Name,
 		kvStore,
 	}, nil
 }
 
-func (stream *Jetstream) Connect(triggerName string, dependencyExpression string, deps []Dependency) (TriggerConnection, error) {
+func (stream *SensorJetstream) Connect(triggerName string, dependencyExpression string, deps []eventbuscommon.Dependency) (eventbuscommon.TriggerConnection, error) {
 	// Generate clientID with hash code
 	hashKey := fmt.Sprintf("%s-%s", stream.sensorName, triggerName)
