+++ eventbus/stan/eventsource/stanSourceConn.go
@@ -1,0 +1,20 @@
+package eventsource
+
+import (
+	"context"
+
+	eventbuscommon "github.com/argoproj/argo-events/eventbus/common"
+	stanbase "github.com/argoproj/argo-events/eventbus/stan/base"
+)
+
+type STANSourceConn struct {
+	*stanbase.STANConnection
+	eventSourceName string
+	subject         string
+}
+
+func (n *STANSourceConn) PublishEvent(ctx context.Context,
+	evt eventbuscommon.Event,
+	message []byte) error {
+	return n.Publish(n.subject, message)
+}
