+++ eventbus/stan/sensor/sensorSTAN.go
@@ -1,0 +1,41 @@
+package sensor
+
+import (
+	"fmt"
+	"math/rand"
+	"time"
+
+	"github.com/argoproj/argo-events/common"
+	eventbuscommon "github.com/argoproj/argo-events/eventbus/common"
+	stanbase "github.com/argoproj/argo-events/eventbus/stan/base"
+	"go.uber.org/zap"
+)
+
+type SensorSTAN struct {
+	*stanbase.STAN
+	sensorName string
+}
+
+func NewSensorSTAN(url, clusterID, sensorName string, auth *eventbuscommon.Auth, logger *zap.SugaredLogger) *SensorSTAN {
+	return &SensorSTAN{
+		stanbase.NewSTAN(url, clusterID, auth, logger),
+		sensorName,
+	}
+}
+
+func (n *SensorSTAN) Connect(triggerName string, dependencyExpression string, deps []eventbuscommon.Dependency) (eventbuscommon.TriggerConnection, error) {
+	// Generate clientID with hash code
+	hashKey := fmt.Sprintf("%s-%s-%s", n.sensorName, triggerName, dependencyExpression)
+
+	s1 := rand.NewSource(time.Now().UnixNano())
+	r1 := rand.New(s1)
+	hashVal := common.Hasher(hashKey)
+	clientID := fmt.Sprintf("client-%v-%v", hashVal, r1.Intn(100))
+
+	conn, err := n.MakeConnection(clientID)
+	if err != nil {
+		return nil, err
+	}
+
+	return NewSTANTriggerConn(conn, n.sensorName, triggerName, dependencyExpression, deps), nil
+}
